navigation

+++ HouseCowork/app/src/main/java/com/polly/housecowork/compose/onboarding/OnBoardingNavigation.kt
@@ -12,7 +12,7 @@ import com.polly.housecowork.utils.Screen
 fun NavGraphBuilder.onboardingNavigation(
     navController: NavController,
     startDestination: String,
-    onOnboardingComplete: () -> Unit
+    onOnboardingComplete: (Int) -> Unit
 ) {
     // todo: should be login page when page is designed
     navigation(
@@ -32,6 +32,5 @@ fun NavGraphBuilder.onboardingNavigation(
         }
 
         houseNavigation(navController, onOnboardingComplete)
-
     }
 }
